requirements.txt
numpy
pandas
typing
torch
attrs
scipy